Fixed NavBar routing
Before the navbar was simply deleting what was on the page and redering a file. Now we are actually navigating through the files. I found that I had to create a folder for each react component and changing the file name to page.js. I also had trouble with applying the navbar to each page and the way I do that is to import tha navbar variable from the navbar.js. We should find out how to revert the css from Eideens original navbar to this one.