Added Profile Implementation
I created the options to choose the meal plan. Implemeneted the table of preferneces on how many times a day they are going to eat. Made it so that they can adjust the default. Lastly created a podium of their favorite dining halls. The dining halls are just text right now but we will need to connect that to the backend. I still need to work on creating a profile state so that their preferences are saved.