Added console logging to talk to backend
Added more options to the dining options. Added a reset button that
resets the entries in the table. Created a function that converts
the main data structure to a map of options and their counts. Added
console log commands that will be used to update the backend with the
logged data.