CSS update/ Podium Logic
I created some logic for podium to change the name that's shown on each podium depended on all time visits. I made a mock data called podiumData which is a dictionary that has a value attatched to each restaurant. I also I changed the the css of the podiums but still need to work on it since it looks a bit ugly right now.