Fixed Routing for SignIn and SignOut Functionality
I changed the top level page.js (index page) for
the project. The index page now calls the function
to sign the user in, and reroutes them to their
home page. Also adjusted the routing for the sign
out functionality, but have not figured out yet
how to keep track of the users state throughout the
entire application.