Fixed the SignIn and SignOut
I fixed the config file, and the SignIn and SignOut
components to properly get the current user of the
app and sign them out. We weren't doing this before,
but we need to do this import from the firebase config
file whenever we are using the current instance of
our firebase app.